Correct copyright attribution
"Global Inc" is presumably a placeholder for a demo project. However, the privacy policy in this project is for Google, and of course, https://www.docsy.dev/ itself appears to show "Global Inc" as the copyright, though it should, in actuality, presumably say Google Inc.

I noticed https://example.docsy.dev/ utilizes "The Docsy Authors" as it's copyright attribution though, so presumably that would be the best copyright here as well.