Remove browsers from PostCSS Config
Specifying browsers via this config is not recommended. The best way to provide browsers is a .browserslistrc file in your project root, or by adding a browserslist key to your package.json. (see: https://github.com/postcss/autoprefixer). However, if you don't define the supported browsers the default browsers option will be used (which is a great setup).
REF: https://github.com/browserslist/browserslist#full-list