Update script as described in the review comments
See comments of gwatts. I tested in locally and deployed it for tests under https://tangly-team.bitbucket.io/. The blog part of the site has tags.
The pull request is based on your comment in the review.
I hope it helps the progress of your feature. It is really cool.