Allow setting the color of the root of the sidebar tree separately from the primary color.
Our site currently sets the primary color to white and also the background
of the sidebar to white. Updating our docsy theme makes it so that the root
of the sidebar tree disappears. This allows us to keep our theming and keep
the root of the sidebar a dark contrasting color from the background.