Allow versioned docs to point to the same page in the different versions
If the version_menu_pagelinks config option is true, then link in the drop-down menu
of the versioned docs points to the current page in the other version, instead of the main page.

This can be useful if the document doesn't change much between the different versions.

Note that if the current page doesn't exist in the other version, the link will be broken.