Fix overlapping content for root folder pages
If you write simple .md page in the root content folder, or an _index.md in a top level folder that doesn't map to one of the other layout styles, the content will render underneath the titlebar. This is because the scss expects the `<main>` element to be inside a div with class `td-main`.