Docsy examples: drop private-repo
The purpose of this section, as described in the opening paragraph, is to allow the user to see examples of Docsy in use as well as have access to the source files to see how Docsy is configured. In that spirit, I've dropped the project listed as an example whose repo is private. Looking at the site, it also isn't obvious at all that Docsy is being used.

Thoughts @LisaFC et al.?